+++ src/utest/util.test.ts
@@ -1,0 +1,13 @@
+//import { expect } from 'chai';
+
+describe('util', () => {
+
+    describe.skip('walk suite', () => {
+    })
+
+    describe.skip('parse for lines, by files', () => {
+    })
+
+    describe.skip('find files for tests', () => {
+    })
+})
